ARCHIVE_DISCONNECTED/backend/verifold/visual/glyph_stego_encoder.py
"""
Glyph Steganographic Encoder
=============================

Generates QR-Gs with dual-layer (visible + steganographic) encoding.
Handles tier-based encoding, decoding, and security warnings.
"""

import qrcode
from typing import Dict, List, Any, Optional, Tuple
from dataclasses import dataclass

@dataclass
class GLYMPHData:
    """Represents GLYMPH data structure"""
    visible_layer: str
    hidden_layer: bytes
    tier_level: int
    emotional_entropy: float

class GlyphStegoEncoder:
    """Handles dual-layer QR-G encoding with steganographic capabilities."""

    def __init__(self):
        # TODO: Initialize steganographic parameters
        self.stego_key = None
        self.encoding_matrix = None

    def encode_dual_layer(self, visible_data: str, hidden_data: bytes, tier: int) -> bytes:
        """Encode both visible and steganographic layers in QR-G."""
        import base64
        from PIL import Image
        from io import BytesIO

        qr = qrcode.QRCode(error_correction=qrcode.constants.ERROR_CORRECT_H)
        qr.add_data(visible_data)
        qr.make(fit=True)
        img = qr.make_image(fill="black", back_color="white").convert("RGB")

        pixels = img.load()
        hidden_bits = ''.join(format(byte, '08b') for byte in hidden_data)
        bit_index = 0

        for y in range(img.size[1]):
            for x in range(img.size[0]):
                r, g, b = pixels[x, y]
                if bit_index < len(hidden_bits):
                    r = (r & ~1) | int(hidden_bits[bit_index])
                    bit_index += 1
                pixels[x, y] = (r, g, b)
                if bit_index >= len(hidden_bits):
                    break
            if bit_index >= len(hidden_bits):
                break

        buffered = BytesIO()
        img.save(buffered, format="PNG")
        return buffered.getvalue()

    def decode_visible_layer(self, qr_image: bytes) -> str:
        """Decode visible layer from QR-G image."""
        import cv2
        import numpy as np
        from pyzbar.pyzbar import decode

        img_np = np.asarray(bytearray(qr_image), dtype=np.uint8)
        image = cv2.imdecode(img_np, cv2.IMREAD_COLOR)
        decoded = decode(image)
        return decoded[0].data.decode("utf-8") if decoded else ""

    def decode_hidden_layer(self, qr_image: bytes, stego_key: bytes) -> Optional[bytes]:
        """Decode hidden steganographic layer with proper key."""
        from PIL import Image
        from io import BytesIO

        try:
            img = Image.open(BytesIO(qr_image))
            pixels = img.load()
            hidden_bits = ""

            for y in range(img.size[1]):
                for x in range(img.size[0]):
                    r, g, b = pixels[x, y]
                    hidden_bits += str(r & 1)

            byte_list = [hidden_bits[i:i+8] for i in range(0, len(hidden_bits), 8)]
            decoded_bytes = bytes([int(b, 2) for b in byte_list if len(b) == 8])
            return decoded_bytes
        except Exception:
            return None

    def generate_security_warning(self, tier_level: int, context: str) -> str:
        """Generate appropriate security warning for tier level."""
        if tier_level >= 3:
            return f"⚠️ WARNING: This GLYMPH contains Tier {tier_level} data.\nContext: {context}\nViewer authentication required."
        elif tier_level == 2:
            return f"🔒 Moderate clearance required for this GLYMPH. Context: {context}"
        else:
            return "🟢 Public GLYMPH - no restriction."

    def validate_glyph_integrity(self, qr_image: bytes) -> bool:
        """Validate QR-G integrity and detect tampering."""
        from PIL import Image
        from io import BytesIO
        try:
            img = Image.open(BytesIO(qr_image))
            return img.size[0] > 50 and img.format == "PNG"
        except Exception:
            return False

# TODO: Implement steganographic algorithms
# TODO: Add error correction for hidden layers
# TODO: Create tier-based security warnings
