ARCHIVE_DISCONNECTED/backend/verifold/requirements.txt
# CollapseHash Project Dependencies
# =================================
# 
# Core post-quantum cryptography and hash verification dependencies
# Install with: pip install -r requirements.txt

# Core Dependencies
# -----------------
# Post-quantum cryptography library
liboqs-python>=0.8.0

# Standard cryptographic libraries
cryptography>=41.0.0
hashlib2>=1.0.0

# Numerical computing and data handling
numpy>=1.24.0

# JSON and data serialization
jsonschema>=4.17.0

# Command Line Interface
# ----------------------
click>=8.1.0
colorama>=0.4.6

# Web Interface Dependencies
# --------------------------
# Web framework (choose one)
flask>=2.3.0
# OR fastapi>=0.100.0
# OR streamlit>=1.25.0

# Template engine
jinja2>=3.1.0

# Web server
gunicorn>=21.0.0

# QR Code Generation
# ------------------
qrcode[pil]>=7.4.0
pillow>=10.0.0

# Hardware Security Module Support
# ---------------------------------
# YubiHSM support (optional)
yubihsm[http]>=3.0.0

# TPM support (optional - platform specific)
# tpm2-pytss>=2.1.0  # Linux/Windows only

# Testing and Development
# -----------------------
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
# -------------
mkdocs>=1.5.0
mkdocs-material>=9.1.0

# Optional: Jupyter Notebook Support
# -----------------------------------
jupyter>=1.0.0
notebook>=7.0.0

# Optional: Database Support
# --------------------------
# SQLite (built into Python)
# OR PostgreSQL support
# psycopg2-binary>=2.9.0
# OR MongoDB support  
# pymongo>=4.4.0

# Optional: Advanced Analytics
# ----------------------------
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Performance Monitoring
# ---------------------------------
psutil>=5.9.0
memory-profiler>=0.61.0

# Development Tools
# -----------------
pre-commit>=3.3.0
tox>=4.6.0

# Security and Compliance
# -----------------------
bandit>=1.7.0
safety>=2.3.0

# Platform-specific dependencies
# ===============================
# 
# Windows:
# --------
# pywin32>=306  # Windows-specific APIs
# 
# Linux:
# ------
# python-systemd>=235  # systemd integration
# 
# macOS:
# ------
# pyobjc>=9.2  # macOS-specific APIs

# Installation Notes
# ==================
# 
# 1. Install core dependencies:
#    pip install -r requirements.txt
# 
# 2. For development:
#    pip install -r requirements.txt -r requirements-dev.txt
# 
# 3. For production deployment:
#    pip install -r requirements.txt --no-dev
# 
# 4. For hardware security modules:
#    pip install yubihsm[http] tpm2-pytss
# 
# 5. For web interface with all features:
#    pip install flask gunicorn qrcode[pil] pillow
# 
# Platform-specific Installation
# ===============================
# 
# Ubuntu/Debian:
# sudo apt-get install build-essential libssl-dev libffi-dev python3-dev
# pip install -r requirements.txt
# 
# CentOS/RHEL:
# sudo yum install gcc openssl-devel libffi-devel python3-devel
# pip install -r requirements.txt
# 
# macOS:
# brew install openssl libffi
# pip install -r requirements.txt
# 
# Windows:
# # Install Microsoft C++ Build Tools
# pip install -r requirements.txt

# Version Compatibility
# ======================
# Python >= 3.8 required
# Tested on Python 3.8, 3.9, 3.10, 3.11
# 
# Note: Some dependencies may require specific versions
# based on your Python version and platform.
opencv-python
pyzbar
pillow
qrcode
